+++ src/parameters/parameter.hpp
@@ -23,11 +23,11 @@
 #include "../shadertastic_common.hpp"
 #include "../shader/shader.h"
 
-static std::string get_full_param_name_static(const std::string &effect_name, const std::string &param_name) {
+inline static std::string get_full_param_name_static(const std::string &effect_name, const std::string &param_name) {
     return effect_name + '.' + param_name;
 }
 
-static std::string get_full_subparam_name_static(const std::string &effect_name, const std::string &param_name) {
+inline static std::string get_full_subparam_name_static(const std::string &effect_name, const std::string &param_name) {
     return effect_name + '_' + param_name;
 }